Added Motion Effects to Project Card
