+++ src/components/OtherProject.tsx
@@ -1,0 +1,9 @@
+const OtherProject = () => {
+  return (
+    <article>
+      <div></div>
+    </article>
+  );
+};
+
+export default OtherProject;
